Send and handle "nickChanged" message also through signaling
When the HPB is used, if the local participant does not send audio nor
video there is no Peer for the local participant, so it is not possible
to send data channel messages (although they can be received if the
other participant sends audio or video and thus has a Peer object). When
the HPB is not used, if both the local participant and the remote one do
not send audio nor video there is no Peer for their connection, so it is
not possible to send or received data channel messages between those
participants.

The nick was only sent through data channel messages, so in the above
cases the nick was not initially shown and neither updated if it was
changed later. Now a "nickChanged" signaling message is introduced to
ensure that the nick is transmitted even if there is no Peer for a
participant.

Even if the mobile apps do not currently handle the "nickChanged" event
they will just ignore it without breaking havoc.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>